Refactor: open, close issue 개수 구하는 로직 수정
