+++ backend/src/main/java/com/team6/issue_tracker/domain/page/dto/IssuePageRequest.java
@@ -1,7 +1,6 @@
 package com.team6.issue_tracker.domain.page.dto;
 
 import com.team6.issue_tracker.domain.model.Status;
-import com.team6.issue_tracker.domain.issue.domain.IssueFilter;
 import lombok.Data;
 
 import java.util.List;
